chore: update global rules
