+++ frontend/src/components/SensitiveData/DataFeatureView.vue
@@ -1,0 +1,2 @@
+<template></template>
+<script lang="ts" setup></script>
