+++ frontend/src/locales/es-ES.json
@@ -618,9 +618,13 @@
       "sensitive-column-list": "Lista de columnas confidenciales",
       "global-masking-rule": "Regla de enmascaramiento global",
       "data-feature": "Característica de datos",
-      "add-rule": "Agregar regla",
       "grant-access": "Autorizará el acceso",
       "never-expires": "Nunca expira",
+      "global-rules": {
+        "add-rule": "Agregar regla",
+        "re-order": "Reordenar",
+        "delete-rule-tip": "¿Eliminar esta regla?"
+      },
       "column-detail": {
         "masking-setting-for-column": "Configuración de enmascaramiento para \"{columna}\"",
         "access-user-list": "Acceder a la lista de usuarios",
