+++ frontend/src/locales/zh-CN.json
@@ -617,9 +617,13 @@
       "sensitive-column-list": "敏感数据列表",
       "global-masking-rule": "全局脱敏规则",
       "data-feature": "数据特征",
-      "add-rule": "添加规则",
       "grant-access": "授权访问",
       "never-expires": "无过期时间",
+      "global-rules": {
+        "add-rule": "添加规则",
+        "re-order": "排序",
+        "delete-rule-tip": "删除这条规则？"
+      },
       "column-detail": {
         "masking-setting-for-column": "\"{column}\" 脱敏设置",
         "access-user-list": "用户访问列表",
